add form salary report

+++ HRM-Frontend/human-resource-management/src/components/AddCategoryForm/AddBonusForm/AddBonusForm.jsx
@@ -20,7 +20,7 @@ const schema = yup.object({
 });
 
 function AddBonusForm(props) {
-  const { error, success } = useToast();
+  const { error, success, warn } = useToast();
 
   let { match, history } = props;
   let { id } = match.params;
@@ -55,6 +55,18 @@ function AddBonusForm(props) {
     };
     fetchBonusCategory();
   }, []);
+
+  useEffect(() => {
+    //Hàm đặt tên cho trang
+    const titlePage = () => {
+      if (dataDetailDMKT.length !== 0) {
+        document.title = `Thay đổi danh mục ${dataDetailDMKT.tenDanhMuc}`;
+      } else if (id === undefined) {
+        document.title = `Tạo danh mục khen thưởng mới`;
+      }
+    };
+    titlePage();
+  }, [dataDetailDMKT]);
 
   const {
     register,
@@ -105,23 +117,27 @@ function AddBonusForm(props) {
       }
       history.goBack();
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không thêm hoặc sửa danh mục được ${errors}`);
     }
   };
 
   const handleDelete = async () => {
     try {
-      await DeleteApi.deleteDMKTvKL(id);
-      await ProductApi.PostLS({
-        tenTaiKhoan: decoded.userName,
-        thaoTac: `Xóa danh mục khen thưởng: ${dataDetailDMKT.tenDanhMuc}`,
-        maNhanVien: decoded.id,
-        tenNhanVien: decoded.givenName,
-      });
-      success("Xoá danh mục khen thưởng thành công");
-      history.goBack();
+      if (dataDetailDMKT.trangThai === "Chưa sử dụng") {
+        await DeleteApi.deleteDMKTvKL(id);
+        await ProductApi.PostLS({
+          tenTaiKhoan: decoded.userName,
+          thaoTac: `Xóa danh mục khen thưởng: ${dataDetailDMKT.tenDanhMuc}`,
+          maNhanVien: decoded.id,
+          tenNhanVien: decoded.givenName,
+        });
+        success("Xoá danh mục khen thưởng thành công");
+        history.goBack();
+      } else {
+        warn(`Danh mục đang được sử dụng`);
+      }
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không xóa được danh mục ${errors}`);
     }
   };
