+++ HRM-Frontend/human-resource-management/src/components/AddCategoryForm/AddDepartmentForm/AddDepartmentForm.jsx
@@ -18,7 +18,7 @@ const schema = yup.object({
     .required("Tên phòng ban không được bỏ trống."),
 });
 function AddDepartmentForm(props) {
-  const { error, success } = useToast();
+  const { error, success, warn } = useToast();
 
   let { match, history } = props;
   let { id } = match.params;
@@ -54,6 +54,18 @@ function AddDepartmentForm(props) {
     };
     fetchDepartmentCategory();
   }, []);
+
+  useEffect(() => {
+    //Hàm đặt tên cho trang
+    const titlePage = () => {
+      if (dataDetailDMPB.length !== 0) {
+        document.title = `Thay đổi danh mục ${dataDetailDMPB.tenPhongBan}`;
+      } else if (id === undefined) {
+        document.title = `Tạo danh mục phòng ban mới`;
+      }
+    };
+    titlePage();
+  }, [dataDetailDMPB]);
 
   useEffect(() => {
     const handleDepartmentId = async () => {
@@ -136,24 +148,27 @@ function AddDepartmentForm(props) {
       }
       history.goBack();
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không thêm hoặc sửa danh mục được ${errors}`);
     }
   };
 
   const handleDelete = async () => {
     try {
-      await DeleteApi.deleteDMPB(id);
-      await ProductApi.PostLS({
-        tenTaiKhoan: decoded.userName,
-        thaoTac: `Xóa phòng ban: ${dataDetailDMPB.tenPhongBan}`,
-        maNhanVien: decoded.id,
-        tenNhanVien: decoded.givenName,
-      });
-      success("Xoá phòng ban thành công");
-
-      history.goBack();
+      if (dataDetailDMPB.trangThai === "Chưa sử dụng") {
+        await DeleteApi.deleteDMPB(id);
+        await ProductApi.PostLS({
+          tenTaiKhoan: decoded.userName,
+          thaoTac: `Xóa phòng ban: ${dataDetailDMPB.tenPhongBan}`,
+          maNhanVien: decoded.id,
+          tenNhanVien: decoded.givenName,
+        });
+        success("Xoá phòng ban thành công");
+        history.goBack();
+      } else {
+        warn(`Danh mục đang được sử dụng`);
+      }
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không xóa được danh mục ${errors}`);
     }
   };
