+++ HRM-Frontend/human-resource-management/src/components/AddCategoryForm/AddNationForm/AddNationForm.jsx
@@ -10,6 +10,7 @@ import DeleteApi from "../../../api/deleteAPI";
 import Dialog from "../../Dialog/Dialog";
 import { useToast } from "../../Toast/Toast";
 import jwt_decode from "jwt-decode";
+import { useDocumentTitle } from "../../../hook/TitleDocument";
 
 const dontAllowOnlySpace = /^\s*\S.*$/g;
 const schema = yup.object({
@@ -21,12 +22,14 @@ const schema = yup.object({
 });
 
 function AddNationForm(props) {
-  const { error, success } = useToast();
+  const { error, success, warn } = useToast();
   let { match, history } = props;
   let { id } = match.params;
 
   const token = sessionStorage.getItem("resultObj");
   const decoded = jwt_decode(token);
+
+  useDocumentTitle("Danh mục dân tộc");
 
   const [dataDetailDMDT, setdataDetailDMDT] = useState([]);
   const [showDialog, setShowDialog] = useState(false);
@@ -43,7 +46,7 @@ function AddNationForm(props) {
   };
 
   useEffect(() => {
-    const fetchNvList = async () => {
+    const fetchNationCategory = async () => {
       try {
         if (id !== undefined) {
           setDescription("Bạn chắc chắn muốn sửa danh mục dân tộc");
@@ -54,8 +57,20 @@ function AddNationForm(props) {
         console.log("false to fetch nv list: ", error);
       }
     };
-    fetchNvList();
+    fetchNationCategory();
   }, []);
+
+  useEffect(() => {
+    //Hàm đặt tên cho trang
+    const titlePage = () => {
+      if (dataDetailDMDT.length !== 0) {
+        document.title = `Thay đổi danh mục ${dataDetailDMDT.tenDanhMuc}`;
+      } else if (id === undefined) {
+        document.title = `Tạo danh mục dân tộc mới`;
+      }
+    };
+    titlePage();
+  }, [dataDetailDMDT]);
 
   const {
     register,
@@ -106,23 +121,27 @@ function AddNationForm(props) {
       }
       history.goBack();
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không thêm hoặc sửa danh mục được ${errors}`);
     }
   };
 
   const handleDelete = async () => {
     try {
-      await DeleteApi.deleteDMDT(id);
-      await ProductApi.PostLS({
-        tenTaiKhoan: decoded.userName,
-        thaoTac: `Xóa danh mục dân tộc: ${dataDetailDMDT.tenDanhMuc}`,
-        maNhanVien: decoded.id,
-        tenNhanVien: decoded.givenName,
-      });
-      history.goBack();
-      success("Xoá danh mục dân tộc thành công");
+      if (dataDetailDMDT.trangThai === "Chưa sử dụng") {
+        await DeleteApi.deleteDMDT(id);
+        await ProductApi.PostLS({
+          tenTaiKhoan: decoded.userName,
+          thaoTac: `Xóa danh mục dân tộc: ${dataDetailDMDT.tenDanhMuc}`,
+          maNhanVien: decoded.id,
+          tenNhanVien: decoded.givenName,
+        });
+        history.goBack();
+        success("Xoá danh mục dân tộc thành công");
+      } else {
+        warn(`Danh mục đang được sử dụng`);
+      }
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không xóa được danh mục ${errors}`);
     }
   };
