+++ HRM-Frontend/human-resource-management/src/components/AddCategoryForm/AddPositionForm/AddPositionForm.jsx
@@ -20,7 +20,7 @@ const schema = yup.object({
 });
 
 function AddPositionForm(props) {
-  const { error, success } = useToast();
+  const { error, success, warn } = useToast();
 
   let { match, history } = props;
   let { id } = match.params;
@@ -56,6 +56,18 @@ function AddPositionForm(props) {
     };
     fetchPositionCategory();
   }, []);
+
+  useEffect(() => {
+    //Hàm đặt tên cho trang
+    const titlePage = () => {
+      if (dataDetailDMCV.length !== 0) {
+        document.title = `Thay đổi danh mục ${dataDetailDMCV.tenChucVu}`;
+      } else if (id === undefined) {
+        document.title = `Tạo danh mục chức vụ mới`;
+      }
+    };
+    titlePage();
+  }, [dataDetailDMCV]);
 
   useEffect(() => {
     const handlePositionId = async () => {
@@ -138,23 +150,27 @@ function AddPositionForm(props) {
       }
       history.goBack();
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không thêm hoặc sửa danh mục được ${errors}`);
     }
   };
 
   const handleDelete = async () => {
     try {
-      await DeleteApi.deleteDMCV(id);
-      await ProductApi.PostLS({
-        tenTaiKhoan: decoded.userName,
-        thaoTac: `Xóa danh mục chức vụ: ${dataDetailDMCV.tenChucVu}`,
-        maNhanVien: decoded.id,
-        tenNhanVien: decoded.givenName,
-      });
-      success("Xoá danh mục chức vụ thành công");
-      history.goBack();
+      if (dataDetailDMCV.trangThai === "Chưa sử dụng") {
+        await DeleteApi.deleteDMCV(id);
+        await ProductApi.PostLS({
+          tenTaiKhoan: decoded.userName,
+          thaoTac: `Xóa danh mục chức vụ: ${dataDetailDMCV.tenChucVu}`,
+          maNhanVien: decoded.id,
+          tenNhanVien: decoded.givenName,
+        });
+        success("Xoá danh mục chức vụ thành công");
+        history.goBack();
+      } else {
+        warn(`Danh mục đang được sử dụng`);
+      }
     } catch (errors) {
-      error(`Có lỗi xảy ra ${errors}`);
+      error(`Không xóa được danh mục ${errors}`);
     }
   };
